Updated the extension to support NHL fantasy leagues.
The code has been split into three sections: a common section with code that can be shared 100% between the two leagues, an NBA section for NBA-specific code, and an NHL section for NHL-specific code.

A lot of the stuff inside the NBA and NHL folders is also identical. It would be possible to share a lot more code, but since this codebase doesn't use any sort of build or module system, we'll keep it very simple for now.

The build scripts have been extended to create extension outputs for both league types. Since the folder structure is more complex now, it's now necessary to run the command yarn unpacked to create nba/nhl folders for local testing.